add devregs, which serves #R/registers to allow full access to the (memory mapped) register space from e.g. limbo.  the functions used to access the registers are in regs.c, don't use them for normal register access.
the mkfile and sds (kernel config) file have been changed to include these files.

dsinit.b now binds #R after /dev.

inferno-ds/root/ is introduced.  it contains a inferno directory hierarchy with a appl, currently only with appl/nds, specially for nds programs.  copy it to the root of your inferno-os checkout.

the first limbo programs are:
- touchprint.b (sent on the mailing list earlier, it just prints mouse events).
- memfiddle.b.

about memfiddle:

it allows playing with the registers (reading, changing bits, writing) using the touch screen.  four bytes can be shown at a time, each bit is a button.  bits can be flipped by clicking on them.  all bits can be zeroed, `oned', inverted at a time.  or a new value can be read, or written.  a listing of registers is kept.  new addresses can be added to the list by increasing the `current address', the byte/short/long-form can be set for the address, and you can cycle through known addresses with the `k' button.

a listing of well-known registers is given in the files /lib/registers/.  each `subject' has a file, e.g. `video'.  the listings are not complete, feel free to fill in the blanks.  the files are included in the sds kernel and are loaded by memfiddle.  the program has a button `K' to load the next file in the list.

a short description of the buttons presented by memfiddle are printed as comment at the top of the memfiddle.b file.